+++ CHANGELOG.md
@@ -44,13 +44,30 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ### Updated
 
+- Fixed bug in `Add` feature that had a chained action after `ProjectTo`. They will now be filtered like so:
+  
+  ```c#
+  var {entityNameLowercase}List = await _db.{entity.Plural}
+  	.AsNoTracking()
+   	.FirstOrDefaultAsync({entity.Lambda} => {entity.Lambda}.{primaryKeyPropName} == {entityNameLowercase}.{primaryKeyPropName}, cancellationToken);
+  
+  return _mapper.Map<{readDto}>({entityNameLowercase}List);
+  ```
+  
 - Initial commit will use system git user and email as author.
+
   - **DOCS:** Can be toggled off to use a generic `Craftsman` author if desired using a `UseSystemGitUser` boolean on your Domain Template
+
 - `Id` on `BaseEntity` is sortable and filterable by default
+
 - Minor logging updates for better json formatting and more information in prod
+
 - GET record, PUT, and DELETE all have typed ids (e.g. `{id:guid}`) on their controllers
+
 - `Development` environment uses a connection string to an actual database now, instead of an in memory db. This can easily be spun up with a `docker-compose` for local development
+
 - Environment is now a singular object that will take in values for local environment variables for development and placed in launch settings and your docker compose. When deploying to other environments, you will use these same environment variables, but pass the appropriate value for that env.
+
   - Updated auth properties to be under an `AuthSettings` object like we do for broker settings
   - Removed `ConnectionString` from env. Docker connection will be added automatically in launch settings
   - Removed `EnvironmentName` as it will always be `Development`